+++ common/src/main/java/io/github/mortuusars/scholar/client/resource/BuiltInResourcePacks.java
@@ -1,7 +1,7 @@
 package io.github.mortuusars.scholar.client.resource;
 
-import io.github.mortuusars.scholar.PlatformHelper;
 import io.github.mortuusars.scholar.Scholar;
+import io.github.mortuusars.scholar.integration.Mods;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 
@@ -20,11 +20,18 @@ public class BuiltInResourcePacks {
                 Component.translatable("resourcepack.scholar.chiseled_bookshelf_colored_books.name"),
                 new Activation(ActivationType.DEFAULT_ENABLED)));
 
-        if (PlatformHelper.isModLoading("lolmcbv")) {
+        if (Mods.MCBV.isLoading()) {
             packs.add(new Pack(
                     Scholar.resource("chiseled_bookshelf_colored_books_lolmcbv_compat"),
                     Component.translatable("resourcepack.scholar.chiseled_bookshelf_colored_books_lolmcbv_compat.name"),
-                    new Activation(ActivationType.DEFAULT_ENABLED, ActivationType.DEFAULT_DISABLED)));
+                    new Activation(ActivationType.DEFAULT_ENABLED)));
+        }
+
+        if (Mods.WOODWORKS.isLoading()) {
+            packs.add(new Pack(
+                    Scholar.resource("chiseled_bookshelf_colored_books_abnww_compat"),
+                    Component.translatable("resourcepack.scholar.chiseled_bookshelf_colored_books_abnww_compat.name"),
+                    new Activation(ActivationType.DEFAULT_ENABLED)));
         }
 
         return packs;