+++ common/src/main/java/io/github/mortuusars/scholar/integration/Mods.java
@@ -4,6 +4,7 @@ import io.github.mortuusars.scholar.PlatformHelper;
 
 public class Mods {
     public static final Mod MCBV = new Mod("lolmcbv");
+    public static final Mod WOODWORKS = new Mod("woodworks");
 
     public record Mod(String id) {
         public boolean isLoaded() {